requirements.txt
aniso8601==9.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
flask-restx==0.2.0
Flask-WTF==0.14.3
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mongoengine==0.23.0
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
